+++ src/choiwonbin/report2/Report2_4.java
@@ -1,16 +1,16 @@
 package choiwonbin.report2;
 
-//3-6. 아래는 화씨(Fahrenheit)를 섭씨(Celcius)로 변환하는 코드이다.
-//변환 공식이 'C = 5/9*(F-32)'라고 할 때, 빈 칸에 알맞은 코드를 넣으시오.
-// 단, 변환값은 소수점 셋째자리에서 반올림하며, Math.round() 함수를 사용하지 않고 처리할 것!
+//3-4. 아래의 코드는 사과를 담는데 필요한 바구니(버켓)의 수를 구하는 코드이다.
+//만일 사과의 수가 123개이고 하나의 바구니에는 10개의 사과를 담을 수 있다면, 13개의 바구니가 필요할 것이다.
+//알맞은 코드를 넣으시오.
 
-//예상 결과 : Fahrenheit:100, Celcius:37.78
-class Exercise3_6 {
+//예상 결과 -> 필요한 바구니의 수 :13
+class Exercise3_4 {
     public static void main(String[] args){
-        int fahrenheit = 100;
-        float celcius = (int)((5f/9f*(fahrenheit-32))*100 + 0.5f)/100f;
+        int numOfApples = 123; // 사과의 개수
+        int sizeOfBucket = 10; // 바구니의 크기(바구니에 담을 수 있는 사과의 개수)
+        int numOfBucket = (numOfApples/sizeOfBucket + ((numOfApples%sizeOfBucket) > 0 ? 1 : 0)); // 모든 사과를 담는데 필요한 바구니의 수
 
-        System.out.println("Fahrenheit:"+fahrenheit);
-        System.out.println("Celcius:"+celcius);
+        System.out.println("필요한 바구니의 수 :"+numOfBucket);
     }
 }
