+++ src/choiwonbin/report2/Report2_5.java
@@ -1,0 +1,12 @@
+package choiwonbin.report2;
+
+//3-5. 아래는 변수 num의 값에 따라 '양수', '음수', '0'을 출력하는 코드이다.
+//삼항연산자를 이용해서 빈칸에 알맞은 코드를 넣으시오.
+//Hint : 삼항 연산자를 두 번 사용할 것!
+class Exercise3_5{
+    public static void main(String[] args){
+        int num = 10;
+        System.out.println(num == 0 ? 0 : (num > 0 ? "양수" : "음수"));
+    }
+}
+//예상 결과 : 양수
